+++ R/SPAS.print.model.legacy.R
@@ -1,4 +1,4 @@
-#' Print the results from a fit of a Stratified-Petersen (SP) model when using the TMB optimizer
+#' Print the results from a fit of a Stratified-Petersen (SP) model from the legacy (i.e. pre using TMB) code
 #' 
 #' This function makes a report of the results of the model fitting
 #' .
@@ -19,22 +19,22 @@
 #' conne.data <- as.matrix(read.csv(conne.data.csv, header=FALSE))
 #' close(conne.data.csv)
 #'  
-#' mod1 <- SPAS.fit.model(conne.data, model.id="Pooling rows 1/2, 5/6; pooling columns 5/6",
+#' mod1 <- SPAS.fit.model.legacy(conne.data, model.id="Pooling rows 1/2, 5/6; pooling columns 5/6",
 #'                       row.pool.in=c("12","12","3","4","56","56"),
-#'                       col.pool.in=c(1,2,3,4,56,56))
+#'                       col.pool.in=c(1,2,3,4,56,56),
+#'                       optMethod.control=list(ftol=.0001))
 #'
-#' SPAS.print.model(mod1)
+#' SPAS.print.model.legacy(mod1)
 
 
 
 
-SPAS.print.model = function(x){
+SPAS.print.model.legacy = function(x){
 #
-#  Print out model results from the OPEN SPAS model fit using TMB
+#  Print out model results from the OPEN SPAS model from the legacy code
 #
-#  Check that object is of the new form
-   if( x$version<"SPAS-R 2020-01-01"){
-       stop("Object was created using previous version of SPAS. Use SPAS.print.model.legacy()")
+   if( x$version>= "SPAS-R 2020-01-01"){
+       stop("Object was created using newer version of SPAS. Use SPAS.print.model()")
    }
    cat("Model Name:",x$model.info$model.id,"\n")
    cat("   Date of Fit:",format(x$date,"%Y-%m-%d %H:%M"),"\n")
@@ -48,21 +48,15 @@ SPAS.print.model = function(x){
    cat("Physical pooling  :", x$input$row.physical.pool, "\n")
    cat("Theta pooling     :", x$input$theta.pool,  "\n")
    cat("CJS pooling       :", x$input$CJSpool,     "\n")
-   cat("\n\n")
+   cat("\n")
    
-   cat("Chapman estimator of population size ",round(x$est$N.Chapman),
-       ' (SE ', round(x$se$N.Chapman), " )\n")
-   cat(" \n\n")
-   cat("Raw data AFTER PHYSICAL (but not logical) POOLING \n")
+   cat(" \n")
+   cat("Raw data AFTER POOLING \n")
    print(x$fit.setup$pooldata)
-   cat("\n")
-   cat("Condition number of XX' where X= (physically) pooled matrix is ", x$kappa, "\n")
-   cat("Condition number of XX' after logical pooling                  ", x$kappa.after.lp, "\n")
-   cat("\n")
-   cat("Large value of kappa (>1000) indicate that rows are approximately proportional which is not good\n")
    cat("\n")
    
    cat("  Conditional   Log-Likelihood:",x$model.info$logL.cond,"   ;  np:",x$model.info$np,";  AICc:", x$model.info$AICc,"\n")
+   #cat("  Unconditional Log-Likelihood:",x$model.info$logL.uncond,";  np:",x$model.info$np,";  AICc:", x$model.info$AICc,"\n")
    cat("\n")
 
    cat("  Code/Message from optimization is: ",
@@ -73,24 +67,24 @@ SPAS.print.model = function(x){
    S <- nrow(temp)-1
    T <- ncol(temp)-1
 
-   temp[1:S, 1:T]  <- round(x$est$real$theta,1)
-   temp[1:S,   T+1]<- round(x$est$real$psi,1);                colnames(temp)[T+1] <- "psi"
-   temp <- cbind(temp,round(c(x$est$real$cap,       NA),3));  colnames(temp)[T+2] <- "cap.prob"
-   temp <- cbind(temp,round(c(x$est$real$exp.factor,NA),1));  colnames(temp)[T+3] <- "exp factor"
-   temp[S+1,1:T] <-   round(as.vector((1-x$est$real$cap)/x$est$real$cap) %*% matrix(x$est$real$theta,nrow=S, ncol=T)) ;   rownames(temp)[S+1] <- "est unmarked" 
-   temp<- cbind(temp, round(c(x$est$real$N.stratum, x$est$real$N)))     ; colnames(temp)[T+4] <- "Pop Est"
+   temp[1:S, 1:T]  <- round(x$real$est.indiv$theta,1)
+   temp[1:S,   T+1]<- round(x$real$est.indiv$psi,1);           colnames(temp)[T+1] <- "psi"
+   temp <- cbind(temp,round(c(x$real$est.indiv$cap,NA),3));    colnames(temp)[T+2] <- "cap.prob"
+   temp <- cbind(temp,round(c( (1-x$real$est.indiv$cap)/x$real$est.indiv$cap,NA),1)); colnames(temp)[T+3] <- "exp factor"
+   temp[S+1,1:T] <-   round(as.vector((1-x$real$est.indiv$cap)/x$real$est.indiv$cap) %*% x$real$est.indiv$theta,1)  ;   rownames(temp)[S+1] <- "est unmarked" 
+   temp<- cbind(temp, round(c(x$real$est.indiv$N.stratum, x$real$est.indiv$N)))     ; colnames(temp)[T+4] <- "Pop Est"
    cat("\nEstimates\n")
    print(temp)
    
    # get the se's of the estimates
    cat("\nSE of above estimates\n")
    temp <- x$fit.setup$pooldata
-   temp[1:S, 1:T]  <- round(x$se$real$theta,1)
-   temp[1:S,   T+1]<- round(x$se$real$psi,1);                 colnames(temp)[T+1] <- "psi"
-   temp <- cbind(temp,round(c(x$se$real$cap,NA),3));          colnames(temp)[T+2] <- "cap.prob"
-   temp <- cbind(temp,round(c(x$se$real$exp.factor,NA),1));   colnames(temp)[T+3] <- "exp factor"
+   temp[1:S, 1:T]  <- round(x$real$est.indiv$theta.se,1)
+   temp[1:S,   T+1]<- round(x$real$est.indiv$psi.se,1);           colnames(temp)[T+1] <- "psi"
+   temp <- cbind(temp,round(c(x$real$est.indiv$cap.se,NA),3));    colnames(temp)[T+2] <- "cap.prob"
+   temp <- cbind(temp,round(c(x$real$est.indiv$exp.factor.se,NA),1));   colnames(temp)[T+3] <- "exp factor"
    temp[S+1,1:T] <-   as.vector(rep(NA,T) )                    ;   rownames(temp)[S+1] <- "est unmarked" 
-   temp<- cbind(temp, round(c(x$se$real$N.stratum, x$se$real$N))); colnames(temp)[T+4] <- "Pop Est"
+   temp<- cbind(temp, round(c(x$real$est.indiv$N.stratum.se, x$real$est.indiv$N.se))); colnames(temp)[T+4] <- "Pop Est"
    print(temp)
    
    # goodness of fit statistics