+++ R/roxygentags.R
@@ -1,0 +1,6 @@
+#' Roxygen commands
+#' @keywords internal
+#' @useDynLib SPAS
+dummy <- function(){
+  return(NULL)
+}